Rename 'Gnutix\Application' into 'Gnutix\Kernel'.
